refactor(scan-nh): rename `addon` to `rust-addon`
