--- libs/ballot-interpreter-nh-next/ts/src/addon.node
@@ -1,1 +1,0 @@
-../../build/addon.node
