+++ libs/ballot-interpreter-nh-next/ts/src/rust-addon.d.ts
