+++ libs/ballot-interpreter-nh-next/ts/src/rust-addon.node
@@ -1,0 +1,1 @@
+../../build/rust-addon.node
